[refactor] - 코드정리(가독성)
